add and test prefix3, postfix3
These combinators auto-build a parse tree, similar to chainR3 and chainL3

However, an interesting question was raised: should `new Arg(x)` be used
to wrap values in a node type that clearly marks them as such?
The consequences of doing so is that you get tons of levels of Arg when
a chain* parser finds a single operand (and no operator).
Are those nestings of Arg useful? Probably not. But they do show the
actual parsing that happend, so I'm torn between auto-getting-rid of them
and leaving them in.

For now, we'll try leaving them out.
This means that, since there's no longer an Arg type, `dump` can't be a
polymorphic method, so we'll make it a polymorphic standalone procedure
instead, and have it type-check its argument.